Fix for workspace symbols not opening the file
