+++ plugins/plugin-languageserver/che-plugin-languageserver-ide/src/main/java/org/eclipse/che/plugin/languageserver/ide/navigation/workspace/FindSymbolAction.java
@@ -34,8 +34,6 @@ import org.eclipse.che.ide.api.action.ActionEvent;
 import org.eclipse.che.ide.api.editor.EditorAgent;
 import org.eclipse.che.ide.api.editor.EditorPartPresenter;
 import org.eclipse.che.ide.api.editor.editorconfig.TextEditorConfiguration;
-import org.eclipse.che.ide.api.editor.text.TextPosition;
-import org.eclipse.che.ide.api.editor.text.TextRange;
 import org.eclipse.che.ide.api.editor.texteditor.TextEditor;
 import org.eclipse.che.ide.dto.DtoFactory;
 import org.eclipse.che.ide.filters.FuzzyMatches;
@@ -49,7 +47,6 @@ import org.eclipse.che.plugin.languageserver.ide.service.WorkspaceServiceClient;
 import org.eclipse.che.plugin.languageserver.ide.util.DtoBuildHelper;
 import org.eclipse.che.plugin.languageserver.ide.util.OpenFileInEditorHelper;
 import org.eclipse.lsp4j.Location;
-import org.eclipse.lsp4j.Range;
 import org.eclipse.lsp4j.ServerCapabilities;
 import org.eclipse.lsp4j.SymbolInformation;
 
@@ -171,26 +168,13 @@ public class FindSymbolAction extends AbstractPerspectiveAction
         Location location = element.getLocation();
         if (location != null && location.getUri() != null) {
           String filePath = location.getUri();
-          Range locationRange = location.getRange();
-
-          TextRange range = null;
-          if (locationRange != null) {
-            range =
-                new TextRange(
-                    new TextPosition(
-                        locationRange.getStart().getLine(),
-                        locationRange.getStart().getCharacter()),
-                    new TextPosition(
-                        locationRange.getEnd().getLine(), locationRange.getEnd().getCharacter()));
-          }
           result.add(
               new SymbolEntry(
                   element.getName(),
                   "",
                   filePath,
-                  filePath,
+                  location,
                   symbolKindHelper.from(element.getKind()),
-                  range,
                   symbolKindHelper.getIcon(element.getKind()),
                   editorHelper,
                   matches));
