+++ plugins/plugin-languageserver/che-plugin-languageserver-ide/src/main/java/org/eclipse/che/plugin/languageserver/ide/navigation/workspace/SymbolEntry.java
@@ -11,47 +11,40 @@
 package org.eclipse.che.plugin.languageserver.ide.navigation.workspace;
 
 import java.util.List;
-import org.eclipse.che.ide.api.editor.text.TextRange;
 import org.eclipse.che.ide.filters.Match;
-import org.eclipse.che.plugin.languageserver.ide.quickopen.EditorQuickOpenEntry;
+import org.eclipse.che.plugin.languageserver.ide.quickopen.QuickOpenEntry;
 import org.eclipse.che.plugin.languageserver.ide.util.OpenFileInEditorHelper;
+import org.eclipse.lsp4j.Location;
 import org.vectomatic.dom.svg.ui.SVGResource;
 
 /** @author Evgen Vidolob */
-class SymbolEntry extends EditorQuickOpenEntry {
+class SymbolEntry extends QuickOpenEntry {
 
   private String name;
   private String parameters;
   private String description;
-  private String filePath;
+  private Location symbolLocation;
   private String type;
-  private TextRange range;
+  private OpenFileInEditorHelper editorHelper;
   private SVGResource icon;
 
   public SymbolEntry(
       String name,
       String parameters,
       String description,
-      String filePath,
+      Location symbolLocation,
       String type,
-      TextRange range,
       SVGResource icon,
       OpenFileInEditorHelper editorHelper,
       List<Match> matches) {
-    super(editorHelper);
     this.name = name;
     this.parameters = parameters;
     this.description = description;
-    this.filePath = filePath;
+    this.editorHelper = editorHelper;
     this.type = type;
-    this.range = range;
     this.icon = icon;
+    this.symbolLocation = symbolLocation;
     setHighlights(matches);
-  }
-
-  @Override
-  protected String getFilePath() {
-    return filePath;
   }
 
   @Override
@@ -73,8 +66,16 @@ class SymbolEntry extends EditorQuickOpenEntry {
     return type;
   }
 
+  public Location getLocation() {
+    return symbolLocation;
+  }
+
   @Override
-  protected TextRange getTextRange() {
-    return range;
+  public boolean run(Mode mode) {
+    if (mode == Mode.OPEN) {
+      editorHelper.openLocation(getLocation());
+      return true;
+    }
+    return false;
   }
 }
